reorganize code to accomodate new field
