+++ frontend/src/pages/ActivityReport/index.js
@@ -7,7 +7,7 @@ import React, {
 } from 'react';
 import PropTypes from 'prop-types';
 import {
-  keyBy, mapValues, startCase, isEqual,
+  keyBy, mapValues, startCase,
 } from 'lodash';
 import { Helmet } from 'react-helmet';
 import ReactRouterPropTypes from 'react-router-prop-types';
@@ -17,7 +17,6 @@ import useDeepCompareEffect from 'use-deep-compare-effect';
 import moment from 'moment';
 import pages from './Pages';
 import Navigator from '../../components/Navigator';
-
 import './index.scss';
 import { NOT_STARTED } from '../../components/Navigator/constants';
 import {
@@ -26,11 +25,10 @@ import {
   LOCAL_STORAGE_DATA_KEY,
   LOCAL_STORAGE_ADDITIONAL_DATA_KEY,
   LOCAL_STORAGE_EDITABLE_KEY,
-  DATE_DISPLAY_FORMAT,
-  DATEPICKER_VALUE_FORMAT,
 } from '../../Constants';
 import { getRegionWithReadWrite } from '../../permissions';
 import useARLocalStorage from '../../hooks/useARLocalStorage';
+import { convertGoalsToFormData, convertReportToFormData, findWhatsChanged } from './formDataHelpers';
 
 import {
   submitReport,
@@ -120,94 +118,6 @@ function setConnectionActiveWithError(e, setConnectionActive) {
   return connection;
 }
 
-/**
- * compares two objects using lodash "isEqual" and returns the difference
- * @param {*} object
- * @param {*} base
- * @returns {} containing any new keys/values
- */
-export const findWhatsChanged = (object, base) => {
-  function reduction(accumulator, current) {
-    if (current === 'startDate' || current === 'endDate') {
-      if (!object[current] || !moment(object[current], 'MM/DD/YYYY').isValid()) {
-        accumulator[current] = null;
-        return accumulator;
-      }
-    }
-
-    if (current === 'creatorRole' && !object[current]) {
-      accumulator[current] = null;
-      return accumulator;
-    }
-
-    // this block intends to fix an issue where multi recipients are removed from a report
-    // after goals have been saved we pass up the removed recipients so that their specific links
-    // to the activity report/goals will be severed on the backend
-    if (current === 'activityRecipients' && !isEqual(base[current], object[current])) {
-      // eslint-disable-next-line max-len
-      const grantIds = object.activityRecipients.map((activityRecipient) => activityRecipient.activityRecipientId);
-      // eslint-disable-next-line max-len
-      accumulator.recipientsWhoHaveGoalsThatShouldBeRemoved = base.activityRecipients.filter((baseData) => (
-        !grantIds.includes(baseData.activityRecipientId)
-      )).map((activityRecipient) => activityRecipient.activityRecipientId);
-
-      // if we change activity recipients we should always ship the goals up as well
-      // we do hit recipients first, so if they were somehow both changed before the API was hit
-      // (unlikely since they are on different parts of the form)
-      // the goals that were changed would overwrite the next line
-      accumulator.goals = base.goals.map((goal) => ({ ...goal, grantIds }));
-    }
-
-    if (!isEqual(base[current], object[current])) {
-      accumulator[current] = object[current];
-    }
-
-    return accumulator;
-  }
-
-  // we sort these so they traverse in a particular order
-  // (ActivityRecipients before goals, in particular)
-  return Object.keys(object).sort().reduce(reduction, {});
-};
-
-export const unflattenResourcesUsed = (array) => {
-  if (!array) {
-    return [];
-  }
-
-  return array.map((value) => ({ value }));
-};
-
-/**
- * Goals created are editable until the report is loaded again. The report used to
- * not update freshly created goals with their DB id once saved, but this caused
- * any additional updates to create a brand new goal instead of updating the old goal.
- * We now use the goal created in the DB. However this means we no longer know if the
- * goal should be editable or not, since it was loaded from the DB. This method takes
- * the list of newly created goals and grabs their names, placed in the `editableGoals`
- * variable. Then all goals returned form the API (the report object passed into this
- * method) have their name compared against the list of fresh goals. The UI then uses
- * the `new` property to determine if a goal should be editable or not.
- * @param {*} report the freshly updated report
- * @returns {function} function that can be used by `setState` to update
- * formData
- */
-export const updateGoals = (report) => (oldFormData) => {
-  const oldGoals = oldFormData.goals || [];
-  const newGoals = report.goals || [];
-
-  const goalsThatUsedToBeNew = oldGoals.filter((goal) => goal.new);
-  const goalsFreshlySavedInDB = goalsThatUsedToBeNew.map((goal) => goal.name);
-  const goals = newGoals.map((goal) => {
-    const goalEditable = goalsFreshlySavedInDB.includes(goal.name);
-    return {
-      ...goal,
-      new: goalEditable,
-    };
-  });
-  return { ...oldFormData, goals, objectivesWithoutGoals: report.objectivesWithoutGoals };
-};
-
 function ActivityReport({
   match, location, region,
 }) {
@@ -260,49 +170,6 @@ function ActivityReport({
     // seeing a save message if they refresh the page after creating a new report.
     history.replace();
   }, [activityReportId, history]);
-
-  const convertGoalsToFormData = (goals, grantIds) => goals.map((goal) => ({ ...goal, grantIds }));
-
-  const convertObjectivesWithoutGoalsToFormData = (
-    objectives, recipientIds,
-  ) => objectives.map((objective) => ({
-    ...objective,
-    recipientIds,
-  }));
-
-  const convertReportToFormData = (fetchedReport) => {
-    let grantIds = [];
-    let otherEntities = [];
-    if (fetchedReport.activityRecipientType === 'recipient' && fetchedReport.activityRecipients) {
-      grantIds = fetchedReport.activityRecipients.map(({ id }) => id);
-    } else {
-      otherEntities = fetchedReport.activityRecipients.map(({ id }) => id);
-    }
-
-    const activityRecipients = fetchedReport.activityRecipients.map((ar) => ({
-      activityRecipientId: ar.id,
-      name: ar.name,
-    }));
-
-    const goals = convertGoalsToFormData(fetchedReport.goalsAndObjectives, grantIds);
-    const objectivesWithoutGoals = convertObjectivesWithoutGoalsToFormData(
-      fetchedReport.objectivesWithoutGoals, otherEntities,
-    );
-    const ECLKCResourcesUsed = unflattenResourcesUsed(fetchedReport.ECLKCResourcesUsed);
-    const nonECLKCResourcesUsed = unflattenResourcesUsed(fetchedReport.nonECLKCResourcesUsed);
-    const endDate = fetchedReport.endDate ? moment(fetchedReport.endDate, DATEPICKER_VALUE_FORMAT).format(DATE_DISPLAY_FORMAT) : '';
-    const startDate = fetchedReport.startDate ? moment(fetchedReport.startDate, DATEPICKER_VALUE_FORMAT).format(DATE_DISPLAY_FORMAT) : '';
-    return {
-      ...fetchedReport,
-      activityRecipients,
-      ECLKCResourcesUsed,
-      nonECLKCResourcesUsed,
-      goals,
-      endDate,
-      startDate,
-      objectivesWithoutGoals,
-    };
-  };
 
   // cleanup local storage if the report has been submitted or approved
   useEffect(() => {
@@ -567,11 +434,14 @@ function ActivityReport({
           reportId.current, { ...updatedFields, version: 2, approverUserIds: approverIds }, {},
         );
 
+        const { goalForEditing, goals } = convertGoalsToFormData(updatedReport.goals);
+
         updateFormData({
           ...updatedReport,
           startDate: moment(updatedReport.startDate, 'YYYY-MM-DD').format('MM/DD/YYYY'),
           endDate: moment(updatedReport.endDate, 'YYYY-MM-DD').format('MM/DD/YYYY'),
-          goals: updatedReport.goalsAndObjectives,
+          goalForEditing,
+          goals,
         }, true);
         setConnectionActive(true);
         updateCreatorRoleWithName(updatedReport.creatorNameWithRole);
